+++ CMakeLists.txt
@@ -13,7 +13,7 @@ add_executable("${TEST_EXECUTABLE}" "tests/main.cpp"
                                     "tests/hopscotch_set_tests.cpp" 
                                     "tests/policy_tests.cpp")
                                     
-target_include_directories("${TEST_EXECUTABLE}" PRIVATE "${Boost_INCLUDE_DIRS}" "src") 
+target_include_directories("${TEST_EXECUTABLE}" PRIVATE "${Boost_INCLUDE_DIRS}" "include") 
 target_link_libraries("${TEST_EXECUTABLE}" ${Boost_LIBRARIES})
 
 if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES "GNU")
